requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
tensorflow==2.15.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0